+++ src/main/java/org/example/Starter.java
@@ -1,18 +1,23 @@
 package org.example;
 
+import org.example.steganography.EmbedText;
+import org.example.steganography.ExtractText;
+
 import java.io.IOException;
 
 public class Starter {
-    private final TextInFromImage text = new TextInFromImage();
+    private final EmbedText embedText = new EmbedText();
+    private final ExtractText extractText = new ExtractText();
+
     public void start() {
 
         String inputImagePath = "src/main/resources/input.png";
         String outputImagePath = "src/main/resources/output.png";
-        String message = "Привіт світ";
+        String message = "Hello world | Привіт світ";
 
         try {
-            text.embedTextInImage(inputImagePath, message, outputImagePath);
-            String extractedText = text.extractTextFromImage(outputImagePath);
+            embedText.embedTextInImage(inputImagePath, message, outputImagePath);
+            String extractedText = extractText.extractTextFromImage(outputImagePath);
             System.out.println("Text in image: " + extractedText);
         } catch (IOException e) {
             e.printStackTrace();
